moved Action classes from view to controller and set main class for project
