+++ src/controller/actions/gameactions/RightArrowAction.java
@@ -1,15 +1,15 @@
-package view.actions.gameactions;
+package controller.actions.gameactions;
 
 import controller.Main;
 import model.Shooter;
 import javax.swing.*;
 import java.awt.event.ActionEvent;
 
-public class UpArrowAction extends AbstractAction
+public class RightArrowAction extends AbstractAction
 {
     @Override
     public void actionPerformed(ActionEvent e) {
         Shooter sh = (Shooter) Main.gameData.friendFigures.get(0);
-        sh.translate(0, -10);
+        sh.translate(10, 0);
     }
 }
